pyproject.toml
[project]
name = "scaffolding-v3"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Jonas Scholz", email = "jonas.scholz123@gmail.com" }]
dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "tqdm>=4.66.4",
    "matplotlib>=3.9.1",
    "hydra-core>=1.3.2",
    "loguru>=0.7.2",
    "wandb>=0.17.5",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.2",
    "pyarrow>=17.0.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "ipykernel>=6.29.5",
    "hydra-optuna-sweeper>=1.2.0",
    "hydra-submitit-launcher>=1.2.0",
    "ruff>=0.9.4",
    "ty>=0.0.1a3",
    "pre-commit>=4.2.0",
    "mlbnb>=0.1.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }

[tool.ty.src]
root = "./src"
